requirements.txt
pyrogram
tgcrypto
python-dotenv
pydantic
pydantic-settings
tortoise-orm[accel]
aiohttp[speedups]
